+++ volux/__main__.py
@@ -34,7 +34,7 @@ from volux.dp_datatools import LivePercentage, clamp
 from volux.VolumeAssistant import VolumeAssistant
 from volux.VolumeBar import VolumeBar
 ### ---- PREFERENCES ---- ###
-program_title = "volux"
+program_title = "Volux"
 program_icon = realpath('icon.png')
 sound_device = "Master"
 default_mixer_name = "Master"
@@ -73,13 +73,7 @@ class BrightnessMode:
             return(VolumeMode)
         else: raise TypeError("_ismuted should be a bool value")
 ### DEFINE STATE MANAGER
-class StateManager:
-    def __init__(self,initial_state):
-        self.state = initial_state
-    def change_state(self,new_state): # request to change states
-        self.state().vacate()
-        new_state().enter()
-        self.state = new_state
+from volux.StateManager import StateManager
 ### CREATE A STATE MANAGER
 sm = StateManager(VolumeMode)
