+++ volux/testfile.txt
@@ -1,0 +1,3 @@
+Hello world!
+New line.
+Another new line.
